More gracefully make calls to Helper()
Instead of checking if the printer adheres to the testing.tt interface
throughout, wrap any non-matching types in a utility with a NOOP
Helper() implementation.

Removes the unnecessary exposure of the provided Printer in the
Asserter.
In the strictest sense, this change isn't backwards compatible, in that
Asserter no longer exposes a Printer field.
If for some obscure reason this affects you, you can replace all calls
to ja.Printer with the type that you passed in to New().